getting owned games from provider
