+++ backend/Cargo.toml
@@ -36,6 +36,8 @@ zip = "0.6.6"
 pelite = "0.10.0"
 tauri-plugin-window-state = "0.1.0"
 webview2-com = "0.27.0"
+async-trait = "0.1.74"
+futures = "0.3.29"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 webkit2gtk = "0.18.2"
