+++ backend/src/providers/manual_provider.rs
@@ -1,3 +1,5 @@
+use async_trait::async_trait;
+
 use super::provider::{
 	ProviderActions,
 	ProviderStatic,
@@ -5,18 +7,24 @@ use super::provider::{
 use crate::{
 	installed_game,
 	mod_loaders::mod_loader,
+	owned_game::OwnedGame,
 	serializable_struct,
 	Result,
 };
 
 serializable_struct!(ManualProvider {});
 
+#[async_trait]
 impl ProviderActions for ManualProvider {
 	fn get_installed_games(
 		&self,
 		mod_loaders: &mod_loader::DataMap,
 	) -> Result<installed_game::Map> {
 		Ok(installed_game::Map::new())
+	}
+
+	async fn get_owned_games(&self) -> Result<Vec<OwnedGame>> {
+		Ok(Vec::new())
 	}
 }
