+++ backend/src/steam/mod.rs
@@ -2,5 +2,4 @@ pub mod appinfo;
 pub mod command;
 pub mod discover_games;
 pub mod id_lists;
-pub mod owned_games;
 pub mod thumbnail;
